requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
drf-yasg
django-ckeditor
